--- app/home/util/expire.py
@@ -1,19 +1,0 @@
-from pytimeparse2 import parse
-
-
-def parse_expire(request) -> str:
-    # Get Expiration from POST or Default
-    expr = ''
-    if request.POST.get('Expires-At') is not None:
-        expr = request.POST['Expires-At'].strip()
-    elif request.POST.get('ExpiresAt') is not None:
-        expr = request.POST['ExpiresAt'].strip()
-    elif request.headers.get('Expires-At') is not None:
-        expr = request.headers['Expires-At'].strip()
-    elif request.headers.get('ExpiresAt') is not None:
-        expr = request.headers['ExpiresAt'].strip()
-    if expr.lower() in ['0', 'never', 'none', 'null']:
-        return ''
-    if parse(expr) is not None:
-        return expr
-    return request.user.default_expire or ''
